moved helpers to MetaStoreUtils
